requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Scraping and Web Automation
selenium==4.17.2
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0

# PDF and Image Processing
pypdf==4.0.1
Pillow==10.2.0

# Environment Variables
python-dotenv==1.0.1

# WebDriver Manager (optional, for automatic driver management)
webdriver-manager==4.0.1
